project-root/pipelines/etl/requirements.txt
pdfplumber
qdrant-client
openai
pydantic
python-dotenv
